+++ CHANGELOG.md
@@ -739,6 +739,10 @@
   not being used.  
   [Rafael Machado](https://github.com/rakaramos)
   [#1116](https://github.com/realm/SwiftLint/issues/1116)
+* Add `educated_sorting` rule that checks the order of array elements if
+  they are already partially sorted.  
+  [Jamie Edge](https://github.com/JamieEdge)
+  [#1038](https://github.com/realm/SwiftLint/issues/1038)
 
 ##### Bug Fixes